HW10 is complete and tested, fixed some bugs with each of the filters, performance seems really good! Also synthesized the HW11 project for HID mouse control.

+++ HW10/firmware/src/DSP.h
@@ -5,6 +5,6 @@ void clear_data_array(signed short * data_array, unsigned char length);
 
 signed short movAvgFilter(signed short * data, unsigned char sampCount);
 signed short finImpRespFilter(signed short * data, unsigned char sampCount);
-signed short infImpRespFilter(signed short * data, signed short priorValue, signed short weightOld, signed short weightNew);
+signed short infImpRespFilter(signed short * data, signed short priorValue, float weightOld, float weightNew);
 
 #endif /* _DSP_H */
